Cleaned out all compile warnings. Added tools::printMatrix() and tools::CreateIdentityMatrix()

+++ resample/resample/constants.cpp
@@ -5,5 +5,5 @@ bool SKIP_VERIFICATION = false;
 bool PRINT_TO_FILE = false;
 const char* RESULTS_FILE = "results.txt";
 unsigned int RUN_COUNT = 1;
-float MAX_DIFF = (float)0.001;
-const float PI = 3.14159265358979323846;
+float MAX_DIFF = 0.001f;
+const float PI = 3.14159265358979323846f;
